refactor: Extract EventsService into Events::Create{Batch}Service
