+++ app/jobs/events/create_job.rb
@@ -5,7 +5,7 @@ module Events
     queue_as :default
 
     def perform(organization, params, timestamp, metadata)
-      result = EventsService.new.create(
+      result = Events::CreateService.new.create(
         organization: organization,
         params: params,
         timestamp: Time.zone.at(timestamp),
